abstracted country resolution out
